feat(fxml): login page default page again
Previously took the login page out of the mix to test the new main-page and have reinstated it as the default page and reworked it to take users to the main-page upon successful login/registration.

Also made some alterations to the codebase to remove dubious comments and align naming conventions.