fix(routines): routines fetchable from database
Fork of Tim's branch which has had the main branch merged into it. Conflicts have all been resolved, also clarified some of the code in the RoutinePreviewController and SearchController.

I have also fixed the implementation for routine searches, and added a method to the SqliteRoutinesDAO for finding routines with a limit which we can use to display routines on the main page.